Reworked P4Command abstraction supporting the three command types
Previous abstraction had a built in assumption that it was always
a table_add, neglecting differences with table_modify and
table_delete.

A bit of a hack but now a P4Command simply has an 'attributes'
attribute, a dictionary with the attribute name as the key,
and strings (table, action), lists (match params, action params),
or ints (handles) as values as appropriate.

The 'command_type' attribute remains to differentiate handling.