prepare for egress_handling_mode implementation
A number of things had to be moved around.  Primarily, supporting
different egress handling modes means we need to be able to
filter relevant rules as we load the virtual device onto the
physical device.  This means we can't push any rules prior to
an 'insert'/'append' call in the case of Chain leases, as we
previously assumed we could do.  This in turn renders
'migrate_virtual_device' irrelevant and requires associated
code to be moved to 'insert'/'append'.  And so forth.

The Chain::insert method does not yet handle different
egress_handling_modes but this commit prepares the way.

Prepare ye the way.