Make packets flow.
MAKE PACKETS FLOW.  THEY FLOW NOW.

Numerous bug fixes contributed to this commit.  Also, this commit
implements basic multicast support -- specifically, broadcasting.
A multicast group is automatically set up at time of Lease
creation, and it includes all of the ports assigned to the
Lease.  The compiler now exports the [MCAST_GRP] token when
the input P4 source includes a modify_field primitive with
bmv2's intrinsic_metadata.mcast_grp as the destination.  This
token is found in the interpretation guide.  When the slice
manager invokes the interpreter to interpret a rule whose
action modifies im.mcast_grp, the result is that the
[MCAST_GRP] token is replaced with the Lease's mcast_grp_id.
Regarding the Lease's mcast_grp_id, this is assigned at time
of Lease creation by the Device class, which is responsible
for ensuring each Lease gets a unique value for mcast_grp_id.

This commit has been tested with the L2 switch.  Packets flow
and latency looks surprisingly good.