Make more packets flow - intelligent handling of egress filtering!
Now the rest our test P4 programs should work - not just the L2
switch.  In fact, this commit was tested vs. the arp proxy and
it worked!  But bigger, the controller now intelligently handles
bmv2 vs non-bmv2 devices when multicasting, a win for demo'ing
this project's support for portable network configuration solutions.