Fix some accounting & migration support issues
Bugs related to misuse of VirtualDevice attributes, esp.
hp4rules and hp4_code_and_rules.  virtualdevice.py has
comments at the top explaining the attributes that otherwise
can be easy to confuse.  The source of confusion is that
we need to track (table, handle) pairs for both the
user-facing perspective as well as the hp4-facing perspective
(a 1-to-many relationship).