Implement new candidate for pcap comparison between sessions
Existing pcap comparison tools assume the pcaps to compare are
taken from the same session, from different interfaces.  This
tools needs to assume different sessions, same interfaces, to
confirm/disprove that the switch response is the same when the
generated traffic is the same.

Challenge is that the total sequence need not be the same, but
the sequence within all packets pertaining to a request from
A to B should be the same between sessions.  The compare_pcaps2.py
script is the candidate to integrate into the compare_pcaps.sh
bash script.