Add partial support for ternary matching on metadata
Looks to be simple to complete... need minor mod of
compiler, and most of the work will be done in the
rule interpreter, but it isn't complicated - the
only difference between exact matching on metadata
and ternary matching on metadata is that the user-
supplied bitmask will replace the FFs in the full
tmeta.data bitmask supplied by the rule template.

Need to re-familiarize w/ bmv2 API for adding ternary
match entries.  But estimate to complete the whole
thing is less than 1 day.