Fix bugs: virt forwarding, test scripting, test 05
Can't do dynamic topology changes with bmv2 devices in mininet
at this time - adding an interface after the simulation has
already started results in a bmv2 switch that doesn't seem to
be capable of passing traffic to/from the new interface.

But this is not an obstacle to the testing we're doing.

This commit enables a long-lived flow to survive adding an L3
router after an existing L2 switch function, see t05.