Fix interpretation of table_set_default when device=='none'
This fix permits interpretation of table_set_default commands
for virtual devices that are not loaded on any physical
device.  Previously, table_set_default commands could only be
handled if the virtual device was loaded on a physical device.

This upgrade should permit uninterrupted traffic processing
when adding a new virtual device to a chain (and it shouldn't
matter where the new vdev is inserted), because we can fully
configure it before chain insertion.