Modify t05 to better achieve stated goals
Topology has been changed, and two slices are now created.

Topology now has 5 hosts on s1, 2 hosts on s2, and s1 still
connected to s2 as before.

Slice A ('jupiter'):
- s1 interfaces connected to:
  -- h1, h2, h3
  -- s2
- all s2 interfaces (connected to h6, h7, and s1)

Slice B ('saturn'):
- s1 interfaces connected to h4 and h5.

This way, we can, say, start an iperf session between h1 and h2,
while we try things out with h3 connecting to h6/h7 as functions
are added/removed.  And we can also have a session between
h4 and h5 and look at the impacts on Slice B while changes are
made to Slice A.