Implement most of remainder of t05
Everything works, though I still need to put an ongoing
bandwidth or latency test that starts at the beginning,
and continues to the end, allowing us to see how
performance and connectivity changes over the course of
t05's many changes.

Also, it would be nice to implement support for changing
the egress handling mode of a virtual device after it
has been deployed - thus we could program the firewall
to act as a bridge, but up unti the end, the egress
handling mode is efalse so it doesn't affect the
egress spec, and then we change the mode to etrue
once all of the other functions are gone, and packets
resume flowing between h1 and h2.

But that's not strictly required.