Remove obsolete notes; add Known Issues inc. headers w/ 2+ offsets
P4 programs may be written such that a header has more than one
potential offset, depending on branches taken in the parser,
but this is not accounted for in hp4.p4, the compiler, nor in
the interpreter.  Adding support would not be difficult but is
not a priority at this time - the fix involves adding the
parse control state as a match parameter to all hp4.p4 tables
supporting virtual ingress pipeline match tables, and
issuing extra rules / handling in the compiler and interpreter.